+++ apps/jaspr_pad/lib/components/playground/output/execution_service.imports.dart
@@ -1,5 +1,11 @@
 // GENERATED FILE, DO NOT MODIFY
 // Generated with jaspr_builder
 
-export '../../../generated/imports/_web.dart' if (dart.library.io) '../../../generated/imports/_stubs.dart'
-    show IFrameElement, window, MessageEvent, IFrameElementOrStubbed, MessageEventOrStubbed;
+export '../../../generated/imports/_web.dart'
+    if (dart.library.io) '../../../generated/imports/_stubs.dart'
+    show
+        IFrameElement,
+        window,
+        MessageEvent,
+        IFrameElementOrStubbed,
+        MessageEventOrStubbed;