+++ apps/jaspr_pad/lib/components/playground/panels/document_panel.dart
@@ -64,7 +64,7 @@ class __DocumentHintMarkdownState extends State<_DocumentHintMarkdown> {
     final apiLink = _dartApiLink(libraryName);
 
     final propagatedType = info.propagatedType;
-    final _mdDocs = '''### `${info.description?.replaceAll('\n', ' ')}`\n\n
+    final mdDocs = '''### `${info.description?.replaceAll('\n', ' ')}`\n\n
 ${hasDartdoc ? "${info.dartdoc}\n\n" : ''}
 ${isVariable ? "$kind\n\n" : ''}
 ${(isVariable && propagatedType != null) ? "**Propagated type:** $propagatedType\n\n" : ''}
@@ -73,7 +73,7 @@ $apiLink\n\n''';
     // Append a 'launch' icon to the 'Open library docs' link.
     //_htmlDocs = _htmlDocs.replaceAll('library docs</a>', "library docs <span class='launch-icon'></span></a>");
 
-    return _mdDocs;
+    return mdDocs;
   }
 
   String _dartApiLink(String? libraryName) {