+++ apps/jaspr_pad/tool/deploy.dart
@@ -3,9 +3,7 @@ import 'dart:io';
 void main() async {
   var service = 'jasprpad';
 
-  changeWorkingDir('../../');
-
-  await run('docker build -t $service -f packages/jaspr_pad/Dockerfile .');
+  await run('docker build -t $service -f apps/jaspr_pad/Dockerfile .');
   await run('docker tag $service gcr.io/jaspr-demo/$service');
   await run('docker push gcr.io/jaspr-demo/$service');
   await run('gcloud run deploy $service --image=gcr.io/jaspr-demo/$service '