+++ examples/flutter_embedding/web/main.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_embedding_demo/components/app.dart';
+import 'package:jaspr/browser.dart';
+
+/// The web entrypoint for the jaspr app.
+void main() {
+  runApp(App());
+}