+++ examples/flutter_embedding/web/main_jaspr.dart
@@ -1,7 +1,0 @@
-import 'package:element_embedding_demo/components/app.dart';
-import 'package:jaspr/browser.dart';
-
-/// The web entrypoint for the jaspr app.
-void main() {
-  runApp(App());
-}