+++ examples/package_riverpod/lib/main.dart
@@ -1,0 +1,8 @@
+import 'package:jaspr/server.dart';
+
+import 'app.dart';
+
+/// Entrypoint for the server
+void main() async {
+  runApp(Document(body: App()));
+}