+++ examples/riverpod_app/lib/main.dart
@@ -1,8 +1,0 @@
-import 'package:jaspr/server.dart';
-
-import 'app.dart';
-
-void main() async {
-  // runs the app component on both server and client
-  runApp(Document(body: App()));
-}