+++ examples/server_custom_middleware/lib/app.dart
@@ -1,0 +1,12 @@
+import 'package:jaspr/jaspr.dart';
+
+/// A basic jaspr component rendering "Hello World".
+class App extends StatelessComponent {
+  @override
+  Iterable<Component> build(BuildContext context) sync* {
+    yield DomComponent(
+      tag: 'p',
+      child: Text('Hello World'),
+    );
+  }
+}