+++ packages/jaspr_cli/lib/src/command_runner.dart
@@ -1,14 +1,20 @@
 import 'dart:io';
+import 'dart:math';
 
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:cli_completion/cli_completion.dart';
 import 'package:mason/mason.dart';
+import 'package:pub_updater/pub_updater.dart';
 
+import 'analytics.dart';
 import 'commands/build_command.dart';
+import 'commands/clean_command.dart';
 import 'commands/create_command.dart';
+import 'commands/doctor_command.dart';
+import 'commands/generate_command.dart';
 import 'commands/serve_command.dart';
-import 'helpers/clean_helper.dart';
+import 'commands/update_command.dart';
 import 'version.dart';
 
 /// The package name.
@@ -24,18 +30,21 @@ class JasprCommandRunner extends CompletionCommandRunner<int> {
       'version',
       abbr: 'v',
       negatable: false,
-      help: 'Print the current version.',
+      help: 'Print the current version info.',
     );
-    argParser.addFlag(
-      'clean',
-      hide: true,
-    );
+    argParser.addFlag('enable-analytics', negatable: false, help: 'Enable anonymous analytics.');
+    argParser.addFlag('disable-analytics', negatable: false, help: 'Disable anonymous analytics.');
     addCommand(CreateCommand());
     addCommand(ServeCommand());
     addCommand(BuildCommand());
+    addCommand(GenerateCommand());
+    addCommand(CleanCommand());
+    addCommand(UpdateCommand());
+    addCommand(DoctorCommand());
   }
 
   final Logger _logger = Logger();
+  final PubUpdater _updater = PubUpdater();
 
   @override
   Future<int> run(Iterable<String> args) async {
@@ -70,16 +79,52 @@ class JasprCommandRunner extends CompletionCommandRunner<int> {
     }
 
     int? exitCode = ExitCode.unavailable.code;
-    if (topLevelResults['version'] == true) {
+    var isVersionCommand = topLevelResults['version'] == true;
+    if (isVersionCommand) {
       _logger.info(jasprCliVersion);
       exitCode = ExitCode.success.code;
-    } else if (topLevelResults['clean'] == true) {
-      await cleanProject(_logger);
-      exitCode = ExitCode.success.code;
-    } else {
-      exitCode = await super.runCommand(topLevelResults);
     }
-    //if (topLevelResults.command?.name != 'update') await _checkForUpdates();
+    if (topLevelResults.command?.name != 'update') {
+      await _checkForUpdates();
+    }
+    if (!isVersionCommand) {
+      if (topLevelResults['disable-analytics']) {
+        disableAnalytics(_logger);
+      } else if (topLevelResults['enable-analytics']) {
+        enableAnalytics(_logger);
+      } else {
+        exitCode = await super.runCommand(topLevelResults);
+      }
+    }
     return exitCode;
   }
+
+  Future<void> _checkForUpdates() async {
+    try {
+      final latestVersion = await _updater.getLatestVersion(packageName);
+      final isUpToDate = jasprCliVersion == latestVersion;
+      if (!isUpToDate) {
+        _logger.info(wrapBox(
+            '${lightYellow.wrap('Update available!')} ${lightCyan.wrap(jasprCliVersion)} \u2192 ${lightCyan.wrap(latestVersion)}\n'
+            'Run ${cyan.wrap('$executableName update')} to update'));
+      }
+    } catch (_) {}
+  }
+}
+
+String wrapBox(String message) {
+  var lines = message.split('\n');
+  var lengths = lines.map((l) => l.replaceAll(RegExp('\x1B\\[\\d+m'), '').length).toList();
+  var maxLength = lengths.reduce(max);
+  var buffer = StringBuffer();
+  var hborder = ''.padLeft(maxLength + 8, '═');
+  buffer.write('╔$hborder╗\n');
+  for (var (i, l) in lines.indexed) {
+    var pad = (maxLength + 8 - lengths[i]) / 2;
+    var padL = ''.padLeft(pad.floor());
+    var padR = ''.padLeft(pad.ceil());
+    buffer.write('║$padL$l$padR║\n');
+  }
+  buffer.write('╚$hborder╝');
+  return buffer.toString();
 }