+++ packages/jaspr_router/test/routing/lazy_routes_test.dart
@@ -8,7 +8,7 @@ import '../utils.dart';
 void main() {
   group('router', () {
     setUpAll(() {
-      mockHistory();
+      mockPlatform();
     });
 
     testComponents('should push lazy route', (tester) async {