+++ packages/jaspr_flutter_embed/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.4
+
+- Support `flutter >= 3.13.0`
+
+## 0.1.3
+
+- `jaspr` upgraded to `0.9.0`
+
 ## 0.1.2
 
 - `jaspr` upgraded to `0.8.0`